+++ packages/stream_chat/CHANGELOG.md
@@ -3,6 +3,7 @@
 🐞 Fixed
 
 - [[#1774]](https://github.com/GetStream/stream-chat-flutter/issues/1774) Fixed failed to execute 'close' on 'WebSocket'.
+- [[#2016]](https://github.com/GetStream/stream-chat-flutter/issues/2016) Fix muted channel's unreadCount incorrectly updated.
 
 ## 9.4.0